Begin project cartographer.
This is the first of many commits that will define the cartographer
project. Because of the importance of this first commit, the
CMakeLists.txt file is already to work with multiple source and
header files, and outputs version numbers to config.h.
Though no libraries have been linked yet, it's a pretty solid base.